gitk: add commit menu entry
When working with diverged branches, some patches can appear several time
on different branches without having the need to merge those branches. On
the other hand you may have to port a specific patch on another branch. The
search with a SHA1 cannot be applied here since they would differ.

This patch adds an entry in the main context menu to highlight every
instance of a commit.

Signed-off-by: Raphael Gallais-Pou <rgallaispou@gmail.com>